auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic==1.16.4",
    "bcrypt==4.0.1",
    "fastapi==0.104.1",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary==2.9.7",
    "pydantic-settings>=2.11.0",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "redis>=7.0.1",
    "sqlalchemy==2.0.23",
    "uvicorn==0.24.0",
]
